Change over to Catppuccin from No Clown Fiesta
A new year, so I wanted a bit of a change to what I stare at for
multiple hours of the day.

The main benefit of Catppuccin is it's _very_ popular and so is being
actively worked on all the time.

I'm also now using Jetbrains Mono instead of a combination of Commit
Mono and Caskaydia Cove. I'm finding it much nicer to look at.